Add all functionalities needed, except that there's one memory leak left around ft_new_id.c

+++ ft_itoa_base_u.c
@@ -6,7 +6,7 @@
 /*   By: galves-d <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/13 12:05:54 by galves-d          #+#    #+#             */
-/*   Updated: 2020/02/28 18:38:01 by galves-d         ###   ########.fr       */
+/*   Updated: 2020/02/29 16:44:24 by galves-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,10 +36,10 @@ static char	*lst_to_str(t_list *lst)
 char		*ft_itoa_base_u(unsigned long int nbr, const char *base,
 		int is_signed)
 {
-	int				sign;
-	unsigned int	u_nbr;
-	t_list			*num;
-	t_list			*aux;
+	int					sign;
+	unsigned long int	u_nbr;
+	t_list				*num;
+	t_list				*aux;
 
 	num = NULL;
 	if (ft_strlen(base) < 2)
